ws-java-rnetty module
This is implemented using the same Jackson JSON process as others. This means it builds up the bytes in memory before writing them out.
It could be optimized further but will diverge from the other implementations ... and then also not pass the validate.py script which doesn't work with chunked responses.